Refactor with ignored cycle to see what changed
