+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/BindableEditor.tsx
@@ -4,17 +4,13 @@ import {
   BindableAttrValue,
   PropValueType,
   LiveBinding,
-  ScopeMeta,
   JsRuntime,
+  ScopeMeta,
 } from '@mui/toolpad-core';
 import { WithControlledProp } from '../../../utils/types';
 import { getDefaultControl } from '../../propertyControls';
+// eslint-disable-next-line import/no-cycle
 import { BindingEditor } from '../BindingEditor';
-import {
-  BindingEditorDialog,
-  BindingEditorDialogProps,
-} from '../BindingEditor/BindingEditorDialog';
-import type { PropBindingEditorDialogProps } from '../BindingEditor/PropBindingEditorDialog';
 
 function renderDefaultControl(params: RenderControlParams<any>) {
   const Control = getDefaultControl({ typeDef: params.propType });
@@ -37,9 +33,6 @@ export interface BindableEditorProps<V> extends WithControlledProp<BindableAttrV
   liveBinding?: LiveBinding;
   globalScope?: Record<string, unknown>;
   globalScopeMeta: ScopeMeta;
-  Dialog?:
-    | React.ComponentType<BindingEditorDialogProps<V>>
-    | React.ComponentType<PropBindingEditorDialogProps<V>>;
   sx?: SxProps;
 }
 
@@ -55,7 +48,6 @@ export default function BindableEditor<V>({
   liveBinding,
   globalScope = {},
   globalScopeMeta = {},
-  Dialog = BindingEditorDialog,
   sx,
 }: BindableEditorProps<V>) {
   const handlePropConstChange = React.useCallback(
@@ -96,7 +88,6 @@ export default function BindableEditor<V>({
           disabled={disabled || !bindable}
           hidden={!bindable}
           liveBinding={liveBinding}
-          Dialog={Dialog}
         />
       </React.Fragment>
     </Stack>
